+++ src/app/grade/grade.service.ts
@@ -1,7 +1,11 @@
 import { BadRequestException, Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
+import { paginate } from 'nestjs-typeorm-paginate'
+import { QueryParams } from 'src/base/types/query-params.type'
 import { ERROR_MESSAGE } from 'src/common/error-message'
+import { Course } from 'src/database/entities/course.entity'
 import { Grade } from 'src/database/entities/grade.entity'
+import { applySorting } from 'src/utils/query-builder'
 import { FindOptionsWhere, Repository } from 'typeorm'
 import { SubjectMapGradeService } from '../subject-map-grade/subject-map-grade.service'
 import { CreateGradeInput } from './dto/create-grade.input'
@@ -11,11 +15,20 @@ import { UpdateGradeInput } from './dto/update-grade.input'
 export class GradeService {
   constructor(
     @InjectRepository(Grade) private gradeRepository: Repository<Grade>,
+    @InjectRepository(Course) private courseRepository: Repository<Course>,
     private subjectMapGradeService: SubjectMapGradeService
   ) {}
 
-  findAll() {
-    return this.gradeRepository.find()
+  findAll(queryParams: QueryParams) {
+    const { sorting, pagination } = queryParams
+
+    const queryBuilder = this.gradeRepository.createQueryBuilder()
+
+    if (sorting) {
+      applySorting(queryBuilder, sorting)
+    }
+
+    return paginate(queryBuilder, pagination)
   }
 
   findMany(criteria: FindOptionsWhere<Grade> | FindOptionsWhere<Grade>[]) {
@@ -59,6 +72,10 @@ export class GradeService {
   }
 
   async remove(id: string) {
+    const publishedCourse = await this.courseRepository.findOneBy({ gradeId: id })
+    if (publishedCourse) {
+      throw new BadRequestException(ERROR_MESSAGE.CANNOT_DELETE_GRADE_HAVE_PUBLISHED_COURSE)
+    }
     await this.findOne({ id })
     await this.gradeRepository.delete({ id })
     await this.subjectMapGradeService.remove({ gradeId: id })
